Rename "macosx/" folders to "macos/"
