+++ MANIFEST.in
@@ -10,7 +10,7 @@ recursive-include frescobaldi_app *.dic
 recursive-include frescobaldi_app *.js
 recursive-include frescobaldi_app *.md
 recursive-include linux org.frescobaldi.Frescobaldi* Makefile
-recursive-include macosx *.svg *.icns *.py *.strings *.sh *.png *.json *.diff
+recursive-include macos *.svg *.icns *.py *.strings *.sh *.png *.json *.diff
 recursive-include windows *.bmp *.py
 recursive-include i18n *.pot *.po Makefile
 global-exclude *~
