+++ frescobaldi_app/macos/file_open_eventhandler.py
