+++ frescobaldi_app/macos/globalmenu.py
