+++ frescobaldi_app/macos/icon_drag_eventhandler.py
