+++ frescobaldi_app/macos/setup.py
