+++ macos/app_resources/InfoPlist.strings
