+++ macos/appdmg/appdmg.json
