+++ macos/icons/frescobaldi.icns
