+++ macos/patch-PyQt5-uic.diff
